+++ src/main/java/io/github/devatherock/emailsender/exception/ApiErrorResponse.java
@@ -1,0 +1,17 @@
+package io.github.devatherock.emailsender.exception;
+
+import java.util.List;
+
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+import lombok.Setter;
+
+/**
+ * Class representing an error response
+ */
+@Getter
+@Setter
+@AllArgsConstructor
+public class ApiErrorResponse {
+    private List<ApiError> errors;
+}